+++ integrations/amazon_bedrock/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [integrations/amazon_bedrock-v3.11.0] - 2025-08-21
+
+### 🚀 Features
+
+- Add `AmazonBedrockDocumentImageEmbedder` component (#2185)
+
 ## [integrations/amazon_bedrock-v3.10.0] - 2025-08-06
 
 ### 🚀 Features